Use glushort for indices to increase vertex count for one model. Now really works=)

+++ EVil Engine/EVEGraphMesh.h
@@ -21,7 +21,7 @@
 + (EVEGraphMesh *)meshWithName:(NSString *)name
                    vertices:(EVEVertexStruct *)vertices
                 vertexCount:(int)vcount
-                    indices:(GLubyte *)indices
+                    indices:(GLushort *)indices
                  indexCount:(int)icount;
 
 @end
